Node class
Added a Node class. Also, figured out how to call methods from one class into another and how to create objects that represent the classes outside. Added a method to broadcast the blockchain, to add neighbours. The resolver is moved to the Node class. It wasn't easy to figure out how to use stuff from one class to another, but now it will be more easy.