Resolver app route
Reworked the resolver method. Added a /resolve_conflicts route to the app. Currently works on a preset number of nodes, all on my system. Will work on adding a Node class, if it's possible. The new links are http://localhost:5000/resolve_conflicts and http://localhost:5000/get_chains is the old link, but I didn't add it last time.