First working blockchain model
After a whole day of working on it, I managed to test the current blockchain with a SQL file, to change it up, added a hasher function, which has the job of hashing the data and creating tensors out of it. It's still unclear if they will be useful, but the program's able to run and mine a block without breaking.